Added the Home page and the logo.
Fixed rendering problems of the graph: show a loading banner until data are ready to render.

The slowdown during graph rendering is due to the number of nodes and links to render. To overcome the problem, it was decided not to visualize the links of multi-input clusters with more than a certain number of addresses (40 in the code).